renamed files/folders, fixed bug when creating new account
