+++ packages/editor/src/components/projectEditor/editors/deployContractModal/DeployContractModal.tsx
@@ -19,7 +19,7 @@ import classNames from 'classnames';
 import style from './style.less';
 import { ConstructorArgumentsList } from './constructorArgumentsList';
 import { ConstructorArgumentsHeader } from './constructorArgumentsHeader';
-import { ContractArgTypes, IProjectItem, IContractArgData, IContractConfiguration } from '../../../../models';
+import { IProjectItem, IContractArgData, IContractConfiguration } from '../../../../models';
 import { IAccount } from '../../../../models/state';
 import { Modal, ModalHeader, Tooltip } from '../../../common';
 import { IconHelp } from '../../../icons';
@@ -40,7 +40,7 @@ interface IState {
     newContractConfig: IContractConfiguration;
 }
 
-export default class ContractConfigModal extends Component<IProps, IState> {
+export default class DeployContractModal extends Component<IProps, IState> {
 
     state = {
         newContractConfig: this.props.selectedContract.config,
